pyproject.toml
[tool.poetry]
name = "hydroml"
version = "0.1.0"
description = "A machine learning package for hydrological predictions"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "hydroml"}]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
numpy = "^1.24.0"
pandas = "^2.0.0"
torch = "^2.0.0"
pyyaml = "^6.0.2"
pytorch-lightning = "^2.4.0"
xarray = "^2024.11.0"
netcdf4 = "^1.7.2"
geopandas = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
isort = "^5.12.0"
jupyter = "^1.1.1"
matplotlib = "^3.9.3"
tensorboard = "^2.18.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 